+++ packages/core/src/components/go-card/go-card.scss
@@ -124,7 +124,6 @@ go-card {
     flex: 1;
     flex-direction: column;
     gap: var(--go-card-inner-gap);
-    height: 100%;
     .card-title-section {
       .card-title {
         font-size: var(--go-card-title-font-size);
