+++ packages/react/tsconfig.json
@@ -24,7 +24,6 @@
     ],
     "exclude": [
       "node_modules",
-      "../../node_modules"
     ],
     "compileOnSave": false,
     "buildOnSave": false
