+++ templates/react-vite/.gitignore
