+++ templates/react-vite/index.html
