+++ templates/react-vite/package.json
@@ -1,10 +1,9 @@
 {
-  "name": "vite-react-starter",
-  "private": true,
+  "name": "@go-ui/react-vite-template",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "start": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
