+++ templates/react-vite/vite.config.js
