--- templates/vite-react-starter/src/App.jsx
@@ -1,33 +1,0 @@
-import { useState } from 'react';
-import { GoLink, GoButton } from '@go-ui/react';
-import reactLogo from './assets/react.svg';
-import './App.css';
-
-function App() {
-  const [count, setCount] = useState(0);
-
-  return (
-    <div className="App">
-      <div>
-        <GoLink href="https://vitejs.dev" target="_blank">
-          <img src="/vite.svg" className="logo" alt="Vite logo" />
-        </GoLink>
-        <a href="https://reactjs.org" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <GoButton variant="primary" onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </GoButton>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">Click on the Vite and React logos to learn more</p>
-    </div>
-  );
-}
-
-export default App;
